fix errors and devide into files
