+++ readStream.js
@@ -2,21 +2,20 @@ const fs = require('fs')
 const path = require('path')
 const { input } = require("./config");
 
-
-const exsistFile = fs.existsSync(path.resolve(input))
 let Rstream
 
 if(input) {
+    const exsistFile = fs.existsSync(path.resolve(input))
    if(exsistFile) {
         Rstream = fs.createReadStream(path.resolve(input), 'utf-8')
    }
    else {
     process.stderr.write("File not found");
-    process.exit();
+    process.exit(0);
    }
 }
 else {
-    dataFromReadStream = process.stdin
+    Rstream = process.stdin
 }
 
 module.exports = Rstream
